Trinity: trial mcp-doctor + wire tick MCP into repo .mcp.json
- trial mcp-doctor: preflight health check for the tick MCP server (handshake +
  tools/list + non-mutating tick_analyze round-trip); throwaway state by
  default, --repo-root . to check real state read-only.
- .mcp.json: register the `tick` server at repo root so Claude Code can
  coordinate via tick_* tools directly.
- mcp/client.js: extract the shared MCP stdio JSON-RPC client (dedupes the
  smoke-test client).

Regression: mcp-smoke 8/8, battery 11/11, confirm-cli-orchestration 7/7.
